Cleaned up overworld tilers to match frosted tiler.
Fixed nil reference errors in InvisibleObjectTiler. Fixed
InvisibleObjectTiler constantly autotiling stuff in editor, instead of
only when the user toggles the 'autotile' setting.

Fixed obstacle-map-indoors to only autotile in the editor when the
'autotile' property is toggled. This prevents unintentional changing of
tilemaps which happened in the previous few commits.

Fixed obstacle-map-indoors variables which should have been declared as
constants.