Extracted overworld corner covers into their own sprite sheet.
Previously, these corner covers were embedded in the midfrost sprite
sheet. This is slightly wasteful because the corner covers are much
smaller than other sprites in the sheet. More importantly, I plan to
expand the midfrost sprite sheet and this is a way to prevent it from
growing beyond 1000x1000 px.