Removed shadow merging viewport logic.
When shadows overlap, they now result in a darker shadow (which is bad.)
The old approach for fixing this involved creating a viewport the size
of the screen, rendering all the shadows to a viewport texture, and
altering the viewport texture's alpha. However this approach does not
work when there are multiple viewports into the same scene, such as what
happens during puzzles.