New wood background. It tiles instead of stretching.
The wood background on the loading screen and puzzle screen stretched before.
This caused it to look strange when the window was stretched very tall.
It now tiles.

While redrawing the wood background to tile, I also changed its
appearance to a simpler two-tone look to match the rest of the game.