Wood background scaling fixes, renames.
Wood background now stretches to the edges of the screen.

Loading screen and puzzle reuse the same scene for the wood background.

Renamed 'wood-backdrop' to less verbose 'wood-bg'. This is also
consistent with the 'Bg' terminology used elsewhere, like 'BgStrobe' and
'BgGlobViewports' in the puzzle scene. Backdrop is used throughout the
game for flat-colored panels used to fade things out.

Fixed PuzzleDemo's 'spawn pickups' key.