New cheats: cambra, moveme
'cambra' cheat affects the career map and lets you move the camera
freely. This is useful for verifying the appearance of different parts
of the map.

'moveme' cheat affects the career map and lets you move the player along
the path. This is useful for verifying that objects on the map don't
interfere with the creatures.