Tutorial framework bugs: TutorialDemo, tutorial duration, target_lines
PuzzleDemo now ignores the start_level so we can test the middle part of
tutorials.

DurationCalculator no longer calculates the duration of tutorials.
Tutorials have many parts so it was inaccurate to calculate 'The player only
has to clear two lines, so it is a short tutorial'

Fixed divide by zero in RankCalculator with very short levels, such as
levels with only one piece. Target_lines was zero (you can't clear a
line with only one piece) resulting in errors trying to figure out how
the player's performance compared with the expected zero lines.

Moved tutorial's 'game_active = true' statement. This was causing
problems in PuzzleDemo -- when testing the last section of a tutorial,
it would display 'Ready?' but never proceed to 'Go!' because it thought
the game was already active.