Refactor: OverworldBg, loaded_environment_path, stool tweaks
Extracted OverworldBg scene.

Removed unnecessary 'loaded_environment_path' property. Resources have a
path which we can reference.

Renamed 'Stool.tscn' to 'IndoorStool.tscn' to clarify that it's for one
kind of stool. The scene cannot be reused because it has a particular
hitbox and sprite offset which is not shared with other stools.