Fixed bug when squishing/rotating simultaneously.
When the player rotated and squished in the same frame, such as when
inserting the final Q piece in a PQV box, the piece could clip into
other pieces. This occurred because the squish target is only calculated
once a frame when the 'piece_disturbed' signal fires. Rotating the piece
would result in the piece squishing to its old squish target before it
was recalculated.

I fixed the bug by changing the logic to squish before rotating, so its
squish target is always accurate. Squishing first has other intuitive
benefits -- namely that the player can press the rotate and squish
buttons simultaneously to squish and rotate a Q piece into a PQV box as
described above.

Added comments to 'schedule_line_clears' function.