Squish move bugs: sfx, mid-air locks
Fixed bug where pieces could lock in mid air with a squish move. This
would occur if the player exhausted all of the piece's lock resets and
then performed a squish move which left the piece in mid-air.

Fixed bug where squish sound effects would sometimes continue after the
squish was completed or cancelled. This was caused by Godot  #37148, a
bug which causes the 'AudioStreamPlayer.stop()' to malfunction if
playback was started in the same frame
(https://github.com/godotengine/godot/issues/37148)