Added current_version property to upgraders.
Upgraders used to compare the newest version to a constant which worked
because SystemSaveUpgrader and PlayerSaveUpgrader were in sync. After
making a change where the player save data uses a different version, the
SystemSaveUpgrader started complaining that it didn't know about version
27bb, even though 27bb was the newest version of the system data.

I've changed upgraders to compare the newest version to a
current_version variable. I've also made this change in
LevelSettingsUpgrader for symmetry, even though it's unnecessary there
because its upgrader is not reused.