Fixed bug where Spacer could get focus
When using the arrow keys on the Chapter Select screen, if you arrowed
off to the right the selection focus would disappear. This is because it
selected an invisible 'Spacer' button in the upper right corner of the
screen. I've changed the Spacer so it is not a button and cannot obtain
focus.

Changed the background of MoodDemo to a more neutral grey. I'm featuring
this in a YouTube video and the old background was kind of ugly.