Bug: successful_levels entries missing from finished_levels
A typo in the LevelHistory script made it so that if the player was
successful in a level (e.g clearing 100 lines in Marathon Normal) the
level was added to LevelHistory.successful_levels, but not to
LevelHistory.finished_levels. I've fixed the bug.

This bug affects player save data, as any finished levels in earlier
versions were not recorded. I've upgraded the player data to 3776 and
added a translation step which fills any gaps in finished_levels with
their entries from successful_levels.