Paranormal tetrominon box building tweaks.
In Paranormal Tetrominon it's possible for a line clear to create a 3x4.
Before, this new box would only be created when the player dropped a
piece which looked strange. It is now created immediately following the
line clear.

In Paranormal Tetrominon it's possible for a piece to build two possible
boxes at once, such as by sliding a second O piece between two horizontal line
pieces. Before, this would always create the top box which was
inconvenient. It now creates the bottom box so the player can continue
working on the top box.