Moved close buttons to corner. Changed shortcut.
The first 'Close' button was made to close the map menu so it had some
weird idiosyncracies. It had a small gap left for the settings menu, and
it closed with the 'Tab' key because this is the way the menu was
opened. None of these idiosyncracies make sense for the
Tutorial/CareerRegion menu, so I've reverted them. The buttons are in
the corner, and the menus close with escape not tab.

Changed focus mode for region menu arrow buttons. Controller players
need a way to change pages. (Maybe this will eventually be done
automatically and they won't need the arrows, but right now they do.)