Removed free roam from main menu.
Most free roam features have been removed, such as chat icons, the logic for
figuring out who you're trying to talk to and how they should answer, the
scrollable level select panel, and a lot of dialog sequences.

The free roam scene itself is still in the 'demo' folder, as it's useful
for testing things like hit boxes and overworld environments.

Renamed 'creature-chat-icon-hook' to 'mouth-hook'. It is now a Node2D
instead of a RemoteTransform2D. We still need it to figure out where the
creature's mouth is, but we do not need its remote transform features.

Renamed 'free-roam-exit' to 'overworld-exit'. These are still around
because I think they help the atmosphere.