Tutorial uses paged panel. Replaced tutorial flag with id.
The scrollable panel is going away soon, it is only used in FreeRoam
mode. There are also not enough tutorials for it to make sense.

Replaced tutorial flag with TUTORIAL_ID constant. There is only one
tutorial region and if there are ever multiple tutorial regions, parts
of the UI will need to be redesigned.