Improved drop signals. Added 'dropped' signal.
Fixed bug where the 'soft_dropped' and 'hard_dropped' signals were
emitted even if the piece's position didn't change. This behavior was
inconsistent with other signals such as 'moved_left' which only fired
when the movement was successful.

Added a new 'dropped' signal which is emitted when the piece falls due
to gravity. There was no signal for this before.