Fixed bug with 'click' sounds at top of playfield.
The rotate/move signals were emitted incorrectly, causing PieceSfx to
emit the 'click' sound at times when it wasn't supposed to. More
specifically, the 'rotate' signal would be emitted between initial
rotation and initial movement, when the piece could still be obstructing
another piece.

Initial rotation and initial movement signals are now emitted after the
piece's movement is complete, and only in the case where the piece does
not cause a top out.