Decorated/animated puzzle messages.
Puzzle messages like 'Good!', 'You win!', 'Ready?' and 'Regret...' are
now decorated with outlines, text movement and sprites.

I experimented with text gradients but canvas item shaders do not play well with text. I tried having Godot's text engine render the outline and applying a shader to the resulting text, but a canvas item shader could not distinguish between the outline and text. I then tried rendering the outline in the shader itself, but the canvas item shader was not called for transparent parts of the text.

https://godotengine.org/qa/34633/how-to-add-gradient-color-to-fonts Many others have had similar problems, and the only solution appears to be to rasterize text into sprites which interferes with localization.

Start button grabs focus.