Tracery. #player.possessive# works.
https://github.com/Althar93/GDTracery

Added GDTracery. We were already using tracery-like syntax for things
like #player# and #sensei#, but the new cutscenes also needed to be able
to say things like "I'm Tom's business partner" which was not possible.

I've modified the tracery implementation in a few ways:

  * I've added localization. English and spanish have their own modifiers for how
    to make a word possessive, and it is easy to add more.
  * I've removed most modifiers, even common ones such as '-ed' and '-s'. We can
    reintroduce these as necessary, but if we have 10 modifiers and 10 languages
    that is 100 functions to write. Fewer is better.
  * Bugfixes, specifically a flaw in GDTracery's 'set_rng' function.
    This flawed setter was unnecessary so I've removed it.

Tracery may be overkill for our game as it includes a lot of features we
are not using, such as randomly generated text. Once our game is more
mature we should consider stripping out the parts we're not using.