Fixed hard-to-read chat themes.
Chat themes with green on black and red on black looked OK, but
blue/purple on black was hard to read. The border_color/accent_color
logic now incorporates a brightness calculation which distinguishes
between the brightness of different hues.

Also fixed chat accents to be more consistent. Bright accents on dark
backgrounds are muted to blend in better. Dark accents on dark backgrounds
are lightened to stand out more.