Fixed 'nametag sides' logic for appearances
The 'nametag sides' logic wasn't working for cutscenes where creatures
enter/exit. This was because it was based on the creature positions, but for
these scenes the creatures didn't exist yet.

I've changed the logic so the nametag sides are recalculated each time a
creature enters or exits the scene.

I also refactored ChatUi so it uses its own ChatTree instance instead of
digging into ChatAdvancer, and changed both to unassign their ChatTree
instance when the conversation ends.