Fixed mismatch between career map/level character
Fixed a bug where the career map would show a certain chef/customer, but
the level would show a different one.

This bug occurred because the 'level posse' for each level was not
stored, but was generated each time it was needed. So when generating
the map it would randomly select one chef, and then when loading the
level it would randomly select a different chef.

The level posse is now generated once and stored, so that it will be
consistent.