U pieces now rotate like J/L pieces.
U pieces used to spawn pointy-side down, and kick the floor twice during
a full 360 degree rotation. They now spawn flat-side down, and only kick
the floor once. This provides a few benefits:

 - The piece movement now resembles J/L pieces so it will be more
   familiar

 - U pieces no longer need to permit extra floor kicks. They used to
   permit 5 floor kicks because otherwise they were too unforgiving at
   20G compared to other pieces

 - U pieces now look like a 'U' in the next piece queue :)

There is one drawback: U pieces and T pieces no longer interlock when hard
dropping without the rotate key. But, this is a rather superficial drawback in
my opinion. We could preserve it if we spawned the U piece in its '2'
orientation, but then the piece rotates off the top of the screen which
is disorienting.