Move O and I pieces left. Flip V piece.
This makes the pieces more consistent in two ways.

Moving the O and I piece to the left gives the game a 'leftward bias' which is
consistent with other similar games, where pieces spawn to the left if they
cannot spawn exactly in the center.

Flipping the V piece makes it so most Turbo Fat pieces stack nicely with
the tetromino on the bottom and pentomino on the top, the only exception
being the U/T pieces which are necessary for a different reason.
Flipping the V piece also makes it it so most pentominos spawn
flat-side-up.