Fix tests: CreatureLibrary, PlayerSaveUpgrader
test-creature-library.gd could fail depending on the order of prior tests. If
the career data was assigned so that the player was earlier in the story, the
sensei's name would be hard-coded to 'Turbo' and fail the replacement test.

test-player-save-upgrader.gd was loading everything as legacy player
data, even later versions which should have been loaded traditionally.