U piece spawns smooth-side up.
I tried getting used to the upside-down U piece but it was too
unintuitive. Every other Turbo Fat piece lets you drop the quadromino
and tetromino on top of each other without rotating, but with U pieces
breaking this rule I kept stacking them upside-down, even after
practicing for a few hours.

To prevent the flipped U piece from immediately rotating off the top of the
screen, I've added 'piece nudges' which will push a piece down if it attempts
to rotate offscreen. I hate the idea of messing with piece rotation in such an
unintuitive way so it only happens in a very specific case:

 - the piece is rotating offscreen
 - the piece was not already offscreen
 - the piece is not triggering any piece kicks

In rotating the U piece, I also moved its shape one cell upward.
Otherwise topping out with a U piece was making it appear on the second
row instead of the first row.