LevelHistory: best_overall_rank, is_level_success
Extracted new LevelHistory.best_overall_rank method. This functionality
was duplicated in a few places which wanted to calculate the player's
best performance on a level.

Extracted new LevelHistory.is_level_success method. This functionalioty
was duplicated in a few places which wanted to determine if the player
had met the success criteria for a level.