Reorganized creatures into story/nonstory.
The old structure of 'primary/secondary' was originally intended to prevent
characters like Skins from appearing as a random customer, but we now
have the 'customer_if' flag for that use case so this directory
structure is unnecessary. I've replaced it with a new grouping of
'story/nonstory'.

This new grouping of 'story/nonstory' helps me because when I go to create new
cutscenes, I know which characters aren't in the game yet. It is an
artificial grouping and has no effect on the game. Customers are loaded
from both directories.

Moved sensei's creature definition out of from its own directory. It didn't
make sense to have a directory just for this one file.