Added puzzle critters, moles
Puzzles now have 'critters', a node which will house different monsters
that interact with the puzzle. This will eventually include animals who do good
things, and imps/demons who do bad things. For now it just includes
moles.

Levels can now add moles with the 'add_mole' and 'advance_moles'
commands. Moles dig for a few turns, and then add a pickup to the playfield. But
they can be interrupted if they are crushed, or if the player clears the row
they are digging.