Added SfxDeconflicter to avoid playing two sounds simultaneously.
This affects cutscenes where multiple creatures might emote the same
face at the same time. Instead of playing a very loud version of the
emote sound, SfxDeconflicter only plays one instance of the sound so the
player hears it at a normal volume.

This node is generic and can apply to any node which uses an
AudioStreamPlayer to play a sound effect.

Fixes #990.