New 'clear_filled_lines' level trigger effect
This effect allows for some creative levels where lines might only be
cleared when the player makes a snack box, or if they drop 3 pieces. But
it is also dangerous, as it allows for levels which delete pieces while
the player is controlling a piece. I've done my best to make this
somewhat resilient -- but it can still result in weird scenarios where
the player's piece is partially deleted. These scenarios seem
unavoidable -- 'not a bug, but a feature' sort of stuff