Apply initial rotation when topping out.
If we don't apply initial rotation, the player can top out in
frustrating scenarios where the piece actually fits if you don't rotate
it. By applying initial rotation, the player can see that they rotated
the piece, and that is why it doesn't fit.

I considered NOT rotating the piece in this scenario and leaving them
alive. This is how it works for piece movement, but for piece rotation
it feels like this would be more frustrating than helpful.

Closes #1697.