Split PuzzleCritterDemo from PuzzleDemo.
I'm about to add a second critter type with several variations i want to
test in the editor -- if we have 2 keys for moles, 6 keys for this new
critter and so on, we will run out of PuzzleDemo keys very quickly.

I've designed PuzzleCritterDemo to toggle 26 different critters (one
for each letter A-Z) with the remaining non-letter keys used to
manipulate those critters.