Carrot improvements: sfx, deconflicting
Added carrot sfx. They make a 'whoosh' sound when they poof in and out,
and a cartoon plane sound when they fly by.

Fixed carrot deconflicting algorithm. This algorithm used to swap
carrots which were too close together, but this often just moved other
carrots which were also too close. The new algorithm is more consistent,
as demonstrated with a unit test.

Fixed remove_carrots() performance degradation. When calling it with a
large number like 999,999, it would still loop 999,999 times even if it
only had to remove 3 carrots. It now terminates the loop prematurely
once the last carrot is removed.

Fixed bug when mixing carrot types, where the smoke size would
increase/decrease. This is because they all shared the same
ParticlesMaterial. The carrot ParticlesMaterial is now local to the
scene.