Fixed bugs when interrupting/restarting tutorials.
When restarting tutorials after the 'Good!' message showed up, the piece
would never appear, softlocking the tutorial. This is because
PieceManager set physics_process to false during the level transition
but did not set it back to true. It now sets physics_process to true
at the start of a new game.

Tutorials would demonstrate weird behavior when interrupted, because
'yield' statements would continue from the previously run tutorial.
These 'yield' statements have been replaced with timers and listeners which are
cleaned up when a new tutorial is started.

The choices which appear when the player is watching tutorial diagrams
were lingering even if the player gave up on the tutorial. This let them
do strange things like choose tutorial options when they weren't playing
anymore. The choices are now hidden when the tutorial ends.

The TutorialModule framework now includes utility methods to make it
easier for tutorials to do something after a delay, or do something
after all messages are shown to the player.

Closes #1180.