Localize chat messages with symbols, 'Hello #player#'
Before, messages like 'I'm #mood#' would fail
to be translated for two reasons.

Firstly, the translation happened when the message was displayed to the user,
after all entities were replaced. So even if we had a translation for
'I'm #mood#', the translation engine would instead search for 'I'm angry'
which was undefined.

Secondly, the translation engine did not translate replaceable entities
like 'angry'.

Both these issues are now fixed.