Fixed blurry icons introduced by godot#64073
https://github.com/godotengine/godot/issues/64073

A regression introduced in Godot 3.5 made the Turbo Fat icon blurry.
Following the suggested advice in this thread, I've ignored the Godot
documentation and generated an icon with the *reversed* size order
(16,32,48,64,128,256) which fixes it.

Presumably this will break at some indeterminate point in the future
when the order is reversed again, so that's exciting.