Added hold piece cheat. Changed limited piece level behavior.
The player can enable a hold piece cheat via the Settings -> Gameplay
menu.

If the player is using Guideline/WASD controls we define a key for them. If the
player is using custom controls, a 'Swap Hold Piece' menu option is
shown and hidden dynamically. We don't want to confuse players by letting them
set a 'Swap Hold Piece' key which doesn't work because they didn't enable the
cheat, which is why we hide it.

Limited piece levels now count the pieces spawned, not the pieces
placed. Adding a 'hold piece' allows the player to skip pieces and not
place them, resulting in unexpected behavior when the player only places
79/80 pieces but has no more pieces to place. So a level that requires
80 pieces can now be cleared by placing 79 pieces and swapping.