Improved keybinds. 'Numpad 2' -> 'Kp 2'
Keybinds now map joypad's L1/R1 buttons to hard drop. This is analogous
to how we remap shift to hard drop as well, and makes sense for the same
reasons -- Turbo Fat encourages chord-based play involving holding 2-3
buttons at once, and it is hard to hold multiple face buttons.

Reordered guideline keybinds; Space, Kp 8, Shift with 'Shift' going
last. Shift is last because it corresponds to the hold key, so I like
the idea of it being removed from the end of the list instead of deleted from
the middle.

Improved WASD keybinds. Up and Space no longer hard drop with the WASD
preset, but I've added KP Enter. Down no longer soft drops. This seems more
consistent with traditional WASD controls for similar games, while enabling
pinky hard drop which I find useful for high speed play.

Renamed 'Numpad 2' to 'Kp 2'. This is a little shorter and more
confusing, but it's consistent with what appears when customizing your
controls. If we want to change this to 'Numpad 2' that's fine but the
two places should be consistent.