Removed top_out_penalty.
The rank algorithm used to decrease the player's rank if they topped
out. The decrease amount was configurable with a 'top_out_penalty'
property. This is no longer the case, the player's score is now just
decreased by ¥100 each time. Sometimes this affects their rank but
sometimes it does not.