Scenes calling get_node() handle nulls better
I modified most of these places to assign the resulting field to 'null'
if the node path is null. There were a few areas (especially relating to
CreatureVisuals) which already had null handling logic which behaved
differently, preserving the field's states if the node path is null. I
kept these the way it is, because that code is especially brittle if the
underlying CreatureVisuals is null anyways. Assigning a null
CreatureVisuals would never do anything helpful.