LineFiller can now leave empty lines.
Instead of filling ALL empty lines, LineFiller now only fills the top
few.

This opens the possibility for levels which fill in pickups every 3
rows, but leave the other 2 rows empty.

This causes a potential negative interaction with levels like 'above and below'
which insert lines in the middle of the playfield as a part of a line
clear. Our new LineFiller algorithm will only look at the top 3 rows, so
it will ignore empty lines in the middle of the playfield. This seems OK
-- if a level like that wanted to have 'filled lines' in the middle of
the playfield it could insert them, instead of inserting empty lines.