Fixed career savescum bug.
The game had a savescum prevention feature which sabotaged the player's
save data before starting a new puzzle. However, this was broken in c081854f632d4e26cc1f51c737ac5ab2c6ab7fe2
when the save data was replaced with an asynchronous system which didn't
save right away.

I've expanded this code so that it is asynchronous as well, and the
player's save data is temporarily sabotaged immediately before being
written to disk, and then restored.