Added 'night mode' visuals for puzzle mode.
This is not accessible in game, but can be toggled through PuzzleDemo by
pressing the 'N' key. It will eventually be enabled by a new vegetable,
the 'night onion'.

This introduces many new visual elements to the Playfield. Some of them
have the potential to introduce lag, such as the starfield particles --
but I have not seen any lag until they appear visible. I considered
dynamically adding/removing these nodes from the scene tree to save on
resources -- if we have 100 level gimmicks, we don't want 500 nodes in
Playfield. But for now there is not a noticable performance hit, and it
is easier to maintain this way since you can see all the nodes in the
godot editor, or turn them on and off as the game is running.