LevelTrigger conditions are now 'any', not 'all'
If a level trigger defines multiple conditions with the same effect, it
used to only fire if ALL conditions were met. It now fires if ANY
conditions are met.

This allows for more complex combinations of triggers, such as triggers
which fire on the 1st, 2nd, 11th, 12th, 21st and 22nd piece, and so on.