Fixed bugs with onion, nighttime when restarting levels.
These bugs especially affected levels that begin at nighttime and toggle to
daytime. When restarting in nighttime, sometimes the onion wouldn't
launch, or it would launch even though it was already in the sky.

The onion now has the ability to skip to the 'floating in the air' mode
which is used when restarting a nighttime level.