Improved comments: customer_index, apache commons
A lot of puzzle code allows a customer_index of -1, where -1 either
means the current customer or a random customer. This behavior was
undocumented and unintuitive. It's now documented.

Removed unnecessarily verbose 'The input arrays are not modified; this
code is adapted from Apache Commons' comments. The input arrays not
being modified seems implicit, and I'd expect a warning if they WERE
modified. The code being adapted from an old version of Apache Commons
is also not helpful information -- the code doesn't really resemble
Apache Commons anymore, and I had to adapt the algorithms anyways.

Fixed some minor comment typos.