Moved/renamed night resources into 'puzzle/night'
I thought about moving them into 'puzzle/critter' because they're
technically all related to the night onion, but decided on this instead.
I think the idea of 'night mode' where the onion is in the sky is
distinct from the idea of the 'onion critter' who dances and flies
around. It can be activated independently in the puzzle demo (although
the onion is still visible in the sky)