Simplified OverworldExit. Removed unused spawns.
OverworldExit is now a sprite. I've removed its hit detection logic and
its CollisionShape2D. None of this is necessary since there is no longer
the functionality of stepping on the exit to transition to a new area.

Removed spawns from FreeRoam environments. These were originally used
because cinematics took place in these environments, but they are now
just used for demos and don't need spawns.