Moved exits to ground layer, instead of z_index=1
OverworldExit used to have a z_index of -1, causing them to be rendered
behind other nodes. This is fine but it contradicts the approach we have
for other floor nodes, which are all placed in their own layer.

I've removed OverworldExit's z_index and moved them into the 'ground'
node.

Added missing LemonFreeRoamEnvironment exits.