Reverted project.godot type changes from Godot's temper tantrum
On March 3rd of this year, Godot spontaneously ignored all class_name
declarations in the project, even when opening older tagged versions of
the project. Eventually it snapped out of it, but during this period it
removed a lot of types in project.godot and replaced them with generic
'Reference' types.

Presumably these scripts corresponded to problematic .gd scripts it
didn't like? I'm not really sure.