Expose PieceManager.states
Sharks add the functionality of being able to eat an entire piece, which
requires us to advance PieceManager's state to the 'wait for the next
piece' state. I've exposed the PieceStates field to enable this, as it's
easy. In the future if this seems like overreach, we could instead
expose a utility method like 'advance_to_wait_for_piece_state()' or
something and keep this field private.