Shark scoring bugfix, increases: 20,20,20 -> 10,20,50
Fixed a bug in PuzzleState which caused all sharks to give $20 when
eating a piece, instead of giving the displayed reward.

Increased shark payouts; even when the medium sharks were bugged to give
out $20, it wasn't enough to compensate for the clumsiness/slowness of
playing with this gimmick. Honestly, we could probably double it again.