Shark animation fixes: '!', 'dance-end'
The 'dance-end' state now shows a '!' bubble instead of a '...' bubble.
This helps express the urgency of 'Hurry, you'll miss it!' better, plus
we're already using the '...' bubble for a different purpose.

The 'dance-end' animations were using incorrect frames, resulting in a
jerky transition from the 'dance' state. Actually, the large shark was
using the wrong frames entirely causing it to briefly become a small
shark.