Mangled dominos now keep original piece names.
Before, mangling a T-piece with small sharks would still report a
'T-spin', but mangling a T-piece with a medium shark would report a
'D-spin' because it is now a domino. ...I've changed it so they will all
report a 'T-Spin' for consistency.

Levels can still be created which have dominos, and those will report a
'D-spin'.