Fixed line_cleared triggers for float_fall levels
Before, float_fall levels would not emit line_cleared triggers properly
because the index of the deleted line did not match the index of the
cleared line. Now, line_cleraed triggers are emitted as long as ANY
lines are being cleared. This means it'll trigger properly for
float_fall, but still won't trigger during a top out.