Avoid adding/removing critters during line clears.
Critters being added/removed during line clears is messy for multiline
clears. For example, critters might spawn as a result of a top line
being cleared, be placed on the bottom line, which is then cleared as
well. This might result in a level which is supposed to spawn 5 critters
only spawning 3, because the other 2 spawned in strange places.

This scenario is now handled by a new "CallQueue". Critters
spawning/advancing is now delayed until after all lines are cleared.
Critters removing is still handled immediately -- this removal occurs in
a specific cell, so if we didn't handle it immediately those coordinates
wouldn't track to the critter's new location without a lot of work.

Fixed some shark/mole comments referencing the wrong critter.