Bug: Sharks crash eating on their last 'patience' turn
Sharks have a 'patience' meter which counts down, and then they
disappear. If a shark is triggered where their patience decreases when a
piece is placed, and then they eat a piece, the game would crash. The
game crashed because the piece would be eaten (counting as a placed
piece) making the shark disappear before they ate.

Sharks now transition to the 'eating state' before the piece is eaten.