Refresh scenes for Godot 3.5.x
I opened all scenes in the editor and saved them to flush out any weird
transient properties, omitted defaults, or differences introduced
between Godot 3.4.x and 3.5.x.

This is in preparation for the Godot 4 upgrade; I'd like to keep these
kinds of changes separate from the changes for the upgrade.