Fix cheeky LineClearer slice call.
LineClearer had a slice call which, for most levels, resulted in
Array.slice(0, -1) which actually has no effect. In Godot 4.x, the slice
function is exclusive instead of inclusive, meaning the code translates
to Array.slice(0, 0) which empties the array, introducing a bug.

It's better just to preempt this bug by avoiding this cheeky code.