Fixed out of bound atlas tiles.
Several atlas TileSets defined regions outside the boundaries of their
corresponding texture. This is seemingly harmless in Godot 3, but
crashes Godot 4 due to Godot issue #67556
https://github.com/godotengine/godot/issues/67556

For most textures, this involved adjusting the region to fit the
corresponding image.

The ubiquitous 'empty-8.png' is not large enough to accommodate any tilemaps,
so I'd increased it to 600x600. This seems silly, and I tried having an empty
image texture but I couldn't find a way to programmatically define an empty
texture in Godot.

'blocks-veggie-cubes-onion.png' was not wide enough, and was missing
tiles for certin vegetables. I've added the missing columns.