Fixed rect_rotation math. Removed Godot #21789 workaround.
The Godot 4 upgrade process replaces 'rect_rotation' with 'rotation',
but rect_rotation is in degrees and rotation is in radians, and the
upgrader does not convert them. I've converted the values.

Removed Godot #21789 workaround to force 'get_class' to return a
specific value. This was implemented because 'if (field is MyClass)' did not
work in Godot 3, but it works in Godot 4 so the workaround is not
necessary.