Refactored 'internal field' workaround
Unlike setget in previous Godot versions, the properties setter and getter are
always called, even when accessed inside the same class (with or without
prefixing with self.). This makes the behavior consistent. Godot's official
guidance is to use another variable for direct access and make the property
code use that name.

We now override the getters to return the internal value as well. This
avoids an edge case where the internal value was set, but the external
value still returned the old value.