Opaque indoor/outdoor shadows.
Godot 3.x has no way to render translucent groups of nodes. We originally
thought it would be feasible to upgrade to Godot 4.x and leverage Godot 4.x
CanvasGroups, but we cannot migrate to Godot 4 due to critical missing features.
More information is available on the godot-4-transition page in the wiki.

Closes #2060.